requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Deprecated==1.3.1
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.121.3
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.10.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.14.0
hdbscan==0.8.41
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.2
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.10.0
libclang==18.1.1
limits==4.2
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.2.1
numpy==2.0.2
open_clip_torch==3.2.0
opt_einsum==3.4.0
optree==0.18.0
packaging==24.2
pillow==11.3.0
pluggy==1.6.0
protobuf==6.33.1
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
timm==1.0.24
tomli==2.4.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==2.0.1
zipp==3.23.0
